requirements.txt
pymongo>=4.8.0
python-dotenv>=1.0.1
requests>=2.32.3
tqdm>=4.66.5